feature: Notification App push

+++ HanRiverUI/HanRiverIOS/HanRiverIOS/Model/HanRiver.swift
@@ -13,7 +13,7 @@ class HanRiverInfo: Identifiable, ObservableObject, Codable {
     var no_ : Int = 0
     var name: String = ""
     //var temperature: Int
-    @Published var temperature: Int = 0
+    @Published var temperature: Double = 0
     
     func encode(to encoder: Encoder) throws {
         //not Use
@@ -25,7 +25,7 @@ class HanRiverInfo: Identifiable, ObservableObject, Codable {
         let container = try decoder.container(keyedBy: CodingKeys.self)
         no_ = try container.decode(Int.self, forKey: .no_)
         name = try container.decode(String.self, forKey: .name)
-        temperature = try container.decode(Int.self, forKey: .temperature)
+        temperature = try container.decode(Double.self, forKey: .temperature)
     }
     
     enum CodingKeys: CodingKey {
